refactor: util renaming
