+++ src/utils/images/clean.js
